moved engine.possibleX to available-command
